refactor: move common classes to separate project
