+++ .github/workflows/dotnet.yml
@@ -25,8 +25,8 @@ jobs:
     - name: Generate NuGet 
       run: dotnet build
     - name: Publish NuGet to nuget.org
-      run: dotnet nuget push src/VirusTotalCore/bin/Release/*.nupkg --api-key ${{ secrets.NUGET_APIKEY }} --source https://api.nuget.org/v3/index.json --skip-duplicate
+      run: dotnet nuget push src/VirusTotalCore*/bin/Release/*.nupkg --api-key ${{ secrets.NUGET_APIKEY }} --source https://api.nuget.org/v3/index.json --skip-duplicate
     - name: Add GitHub as source
       run: dotnet nuget add source --username hunterlan --password ${{ secrets.GITHUB_TOKEN }} --store-password-in-clear-text --name github "https://nuget.pkg.github.com/hunterlan/index.json"
     - name: Publish NuGet to GitHub
-      run: dotnet nuget push src/VirusTotalCore/bin/Release/*.nupkg  --api-key ${{ secrets.NUGET_PUBLISH_TO_GITHUB }} --source "github" --skip-duplicate
+      run: dotnet nuget push src/VirusTotalCore*/bin/Release/*.nupkg  --api-key ${{ secrets.NUGET_PUBLISH_TO_GITHUB }} --source "github" --skip-duplicate
