+++ VirusTotalCore.sln
@@ -4,19 +4,61 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VirusTotalCore", "src\Virus
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VirusTotalCore.Tests", "tests\VirusTotalCore.Tests\VirusTotalCore.Tests.csproj", "{CFA7370B-860E-43F8-89E6-1FAFEFD07174}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{827E0CD3-B72D-47B6-A68D-7590B98EB39B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VirusTotalCore.Common", "src\VirusTotalCore.Common\VirusTotalCore.Common.csproj", "{1BB3E737-DDE4-4424-A0E5-BB8AA050E37A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{C2D5EBFF-CAD9-49D8-8019-D567EDB02D57}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C2D5EBFF-CAD9-49D8-8019-D567EDB02D57}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C2D5EBFF-CAD9-49D8-8019-D567EDB02D57}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{C2D5EBFF-CAD9-49D8-8019-D567EDB02D57}.Debug|x64.Build.0 = Debug|Any CPU
+		{C2D5EBFF-CAD9-49D8-8019-D567EDB02D57}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C2D5EBFF-CAD9-49D8-8019-D567EDB02D57}.Debug|x86.Build.0 = Debug|Any CPU
 		{C2D5EBFF-CAD9-49D8-8019-D567EDB02D57}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C2D5EBFF-CAD9-49D8-8019-D567EDB02D57}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C2D5EBFF-CAD9-49D8-8019-D567EDB02D57}.Release|x64.ActiveCfg = Release|Any CPU
+		{C2D5EBFF-CAD9-49D8-8019-D567EDB02D57}.Release|x64.Build.0 = Release|Any CPU
+		{C2D5EBFF-CAD9-49D8-8019-D567EDB02D57}.Release|x86.ActiveCfg = Release|Any CPU
+		{C2D5EBFF-CAD9-49D8-8019-D567EDB02D57}.Release|x86.Build.0 = Release|Any CPU
 		{CFA7370B-860E-43F8-89E6-1FAFEFD07174}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{CFA7370B-860E-43F8-89E6-1FAFEFD07174}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CFA7370B-860E-43F8-89E6-1FAFEFD07174}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{CFA7370B-860E-43F8-89E6-1FAFEFD07174}.Debug|x64.Build.0 = Debug|Any CPU
+		{CFA7370B-860E-43F8-89E6-1FAFEFD07174}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{CFA7370B-860E-43F8-89E6-1FAFEFD07174}.Debug|x86.Build.0 = Debug|Any CPU
 		{CFA7370B-860E-43F8-89E6-1FAFEFD07174}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CFA7370B-860E-43F8-89E6-1FAFEFD07174}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CFA7370B-860E-43F8-89E6-1FAFEFD07174}.Release|x64.ActiveCfg = Release|Any CPU
+		{CFA7370B-860E-43F8-89E6-1FAFEFD07174}.Release|x64.Build.0 = Release|Any CPU
+		{CFA7370B-860E-43F8-89E6-1FAFEFD07174}.Release|x86.ActiveCfg = Release|Any CPU
+		{CFA7370B-860E-43F8-89E6-1FAFEFD07174}.Release|x86.Build.0 = Release|Any CPU
+		{1BB3E737-DDE4-4424-A0E5-BB8AA050E37A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1BB3E737-DDE4-4424-A0E5-BB8AA050E37A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1BB3E737-DDE4-4424-A0E5-BB8AA050E37A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{1BB3E737-DDE4-4424-A0E5-BB8AA050E37A}.Debug|x64.Build.0 = Debug|Any CPU
+		{1BB3E737-DDE4-4424-A0E5-BB8AA050E37A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{1BB3E737-DDE4-4424-A0E5-BB8AA050E37A}.Debug|x86.Build.0 = Debug|Any CPU
+		{1BB3E737-DDE4-4424-A0E5-BB8AA050E37A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1BB3E737-DDE4-4424-A0E5-BB8AA050E37A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1BB3E737-DDE4-4424-A0E5-BB8AA050E37A}.Release|x64.ActiveCfg = Release|Any CPU
+		{1BB3E737-DDE4-4424-A0E5-BB8AA050E37A}.Release|x64.Build.0 = Release|Any CPU
+		{1BB3E737-DDE4-4424-A0E5-BB8AA050E37A}.Release|x86.ActiveCfg = Release|Any CPU
+		{1BB3E737-DDE4-4424-A0E5-BB8AA050E37A}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{1BB3E737-DDE4-4424-A0E5-BB8AA050E37A} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 	EndGlobalSection
 EndGlobal
