+++ src/VirusTotalCore.Common/Exceptions/AlreadyExistsException.cs
@@ -1,0 +1,3 @@
+namespace VirusTotalCore.Common.Exceptions;
+
+public class AlreadyExistsException(string message) : Exception(message);
