+++ src/VirusTotalCore.Common/Exceptions/AuthenticationRequiredException.cs
@@ -1,4 +1,4 @@
-namespace VirusTotalCore.Exceptions;
+namespace VirusTotalCore.Common.Exceptions;
 /// <summary>
 /// Exception for AuthenticationRequiredException response: "ApiKey is missing"
 /// </summary>
