+++ src/VirusTotalCore.Common/Exceptions/BadRequestException.cs
@@ -1,0 +1,3 @@
+namespace VirusTotalCore.Common.Exceptions;
+
+public class BadRequestException(string message) : Exception(message);
