+++ src/VirusTotalCore.Common/Exceptions/DeadlineExceededException.cs
@@ -1,0 +1,3 @@
+namespace VirusTotalCore.Common.Exceptions;
+
+public class DeadlineExceededException(string message) : Exception(message);
