+++ src/VirusTotalCore.Common/Exceptions/FailedDependencyException.cs
@@ -1,0 +1,3 @@
+namespace VirusTotalCore.Common.Exceptions;
+
+public class FailedDependencyException(string message) : Exception(message);
