+++ src/VirusTotalCore.Common/Exceptions/ForbiddenException.cs
@@ -1,0 +1,3 @@
+namespace VirusTotalCore.Common.Exceptions;
+
+public class ForbiddenException(string message) : Exception(message);
