+++ src/VirusTotalCore.Common/Exceptions/InvalidArgumentException.cs
@@ -1,0 +1,3 @@
+namespace VirusTotalCore.Common.Exceptions;
+
+public class InvalidArgumentException(string message) : Exception(message);
