+++ src/VirusTotalCore.Common/Exceptions/NotAvailableYetException.cs
@@ -1,0 +1,3 @@
+namespace VirusTotalCore.Common.Exceptions;
+
+public class NotAvailableYetException(string message) : Exception(message);
