+++ src/VirusTotalCore.Common/Exceptions/QuotaExceededException.cs
@@ -1,0 +1,3 @@
+namespace VirusTotalCore.Common.Exceptions;
+
+public class QuotaExceededException(string message) : Exception(message);
