+++ src/VirusTotalCore.Common/Exceptions/TooManyRequestsException.cs
@@ -1,0 +1,3 @@
+namespace VirusTotalCore.Common.Exceptions;
+
+public class TooManyRequestsException(string message) : Exception(message);
