+++ src/VirusTotalCore.Common/Exceptions/TransientException.cs
@@ -1,0 +1,3 @@
+namespace VirusTotalCore.Common.Exceptions;
+
+public class TransientException(string message) : Exception(message);
