+++ src/VirusTotalCore.Common/Exceptions/UnselectiveContentQueryException.cs
@@ -1,0 +1,3 @@
+namespace VirusTotalCore.Common.Exceptions;
+
+public class UnselectiveContentQueryException(string message) : Exception(message);
