+++ src/VirusTotalCore.Common/Exceptions/UnsupportedContentQueryException.cs
@@ -1,0 +1,3 @@
+namespace VirusTotalCore.Common.Exceptions;
+
+public class UnsupportedContentQueryException(string message) : Exception(message);
