+++ src/VirusTotalCore.Common/Exceptions/UserNotActiveException.cs
@@ -1,0 +1,3 @@
+namespace VirusTotalCore.Common.Exceptions;
+
+public class UserNotActiveException(string message) : Exception(message);
