+++ src/VirusTotalCore.Common/Exceptions/WrongCredentialsException.cs
@@ -1,4 +1,4 @@
-namespace VirusTotalCore.Exceptions;
+namespace VirusTotalCore.Common.Exceptions;
 /// <summary>
 /// Exception for WrongCredentialsException response: "Wrong API key"
 /// </summary>
