--- src/VirusTotalCore/Exceptions/AlreadyExistsException.cs
@@ -1,3 +1,0 @@
-namespace VirusTotalCore.Exceptions;
-
-public class AlreadyExistsException(string message) : Exception(message);
