--- src/VirusTotalCore/Exceptions/BadRequestException.cs
@@ -1,3 +1,0 @@
-namespace VirusTotalCore.Exceptions;
-
-public class BadRequestException(string message) : Exception(message);
