--- src/VirusTotalCore/Exceptions/DeadlineExceededException.cs
@@ -1,3 +1,0 @@
-namespace VirusTotalCore.Exceptions;
-
-public class DeadlineExceededException(string message) : Exception(message);
