--- src/VirusTotalCore/Exceptions/FailedDependencyException.cs
@@ -1,3 +1,0 @@
-namespace VirusTotalCore.Exceptions;
-
-public class FailedDependencyException(string message) : Exception(message);
