--- src/VirusTotalCore/Exceptions/ForbiddenException.cs
@@ -1,3 +1,0 @@
-namespace VirusTotalCore.Exceptions;
-
-public class ForbiddenException(string message) : Exception(message);
