--- src/VirusTotalCore/Exceptions/InvalidArgumentException.cs
@@ -1,3 +1,0 @@
-namespace VirusTotalCore.Exceptions;
-
-public class InvalidArgumentException(string message) : Exception(message);
