--- src/VirusTotalCore/Exceptions/NotAvailableYetException.cs
@@ -1,3 +1,0 @@
-namespace VirusTotalCore.Exceptions;
-
-public class NotAvailableYetException(string message) : Exception(message);
