--- src/VirusTotalCore/Exceptions/QuotaExceededException.cs
@@ -1,3 +1,0 @@
-namespace VirusTotalCore.Exceptions;
-
-public class QuotaExceededException(string message) : Exception(message);
