--- src/VirusTotalCore/Exceptions/TooManyRequestsException.cs
@@ -1,3 +1,0 @@
-namespace VirusTotalCore.Exceptions;
-
-public class TooManyRequestsException(string message) : Exception(message);
