--- src/VirusTotalCore/Exceptions/TransientException.cs
@@ -1,3 +1,0 @@
-namespace VirusTotalCore.Exceptions;
-
-public class TransientException(string message) : Exception(message);
