--- src/VirusTotalCore/Exceptions/UnselectiveContentQueryException.cs
@@ -1,3 +1,0 @@
-namespace VirusTotalCore.Exceptions;
-
-public class UnselectiveContentQueryException(string message) : Exception(message);
