--- src/VirusTotalCore/Exceptions/UnsupportedContentQueryException.cs
@@ -1,3 +1,0 @@
-namespace VirusTotalCore.Exceptions;
-
-public class UnsupportedContentQueryException(string message) : Exception(message);
