--- src/VirusTotalCore/Exceptions/UserNotActiveException.cs
@@ -1,3 +1,0 @@
-namespace VirusTotalCore.Exceptions;
-
-public class UserNotActiveException(string message) : Exception(message);
