+++ src/VirusTotalCore/Models/Comments/AddComment.cs
@@ -1,5 +1,5 @@
 using System.Diagnostics.CodeAnalysis;
-using VirusTotalCore.Models.Shared;
+using VirusTotalCore.Common.Models.Shared;
 
 namespace VirusTotalCore.Models.Comments;
