+++ src/VirusTotalCore/Models/Comments/CommentData.cs
@@ -1,5 +1,5 @@
 ﻿using System.Text.Json.Serialization;
-using VirusTotalCore.Models.Shared;
+using VirusTotalCore.Common.Models.Shared;
 
 namespace VirusTotalCore.Models.Comments;
