+++ src/VirusTotalCore/Models/Votes/VoteData.cs
@@ -1,5 +1,5 @@
 
-using VirusTotalCore.Models.Shared;
+using VirusTotalCore.Common.Models.Shared;
 
 namespace VirusTotalCore.Models.Votes;
