+++ src/VirusTotalCore/VirusTotalCore.csproj
@@ -19,4 +19,8 @@
     <PackageReference Include="Microsoft.Extensions.Http" Version="8.0.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="../VirusTotalCore.Common/VirusTotalCore.Common.csproj" />
+  </ItemGroup>
+
 </Project>
